requirements.txt
# Jupyter Book and core dependencies
jupyter-book>=1.0.0,<2.0.0
jupytext>=1.16.0,<2.0.0

# Visualization and data handling
matplotlib>=3.8.0,<4.0.0
numpy>=1.26.0,<2.0.0

# Deployment
ghp-import>=2.1.0,<3.0.0

# Configuration and metadata parsing
PyYAML>=6.0.1,<7.0.0
python-frontmatter>=1.1.0,<2.0.0